added default instance type and separated files (#62)
