+++ configuration/provider/aws/ec2/instance/available.go
@@ -1,0 +1,86 @@
+package instance
+
+// Available returns the full list of all available instance types as defined by
+// the constants of this package. Using this medthod ensures that always the
+// same immutable list will be used when referencing all available instance
+// types EC2 provides.
+func Available() []kind {
+	return []kind{
+		TypeM1Small,
+		TypeM1Medium,
+		TypeM1Large,
+		TypeM1XLarge,
+		TypeC1Medium,
+		TypeC1XLarge,
+		TypeCC28XLarge,
+		TypeCG14XLarge,
+		TypeM2XLarge,
+		TypeM22XLarge,
+		TypeM24XLarge,
+		TypeCR18XLarge,
+		TypeI2XLarge,
+		TypeI22XLarge,
+		TypeI24XLarge,
+		TypeI28XLarge,
+		TypeHI14XLarge,
+		TypeHS18XLarge,
+		TypeT1Micro,
+		TypeT2Nano,
+		TypeT2Micro,
+		TypeT2Small,
+		TypeT2Medium,
+		TypeT2Large,
+		TypeT2XLarge,
+		TypeT22XLarge,
+		TypeM4Large,
+		TypeM4XLarge,
+		TypeM42XLarge,
+		TypeM44XLarge,
+		TypeM410XLarge,
+		TypeM416XLarge,
+		TypeM3Medium,
+		TypeM3Large,
+		TypeM3XLarge,
+		TypeM32XLarge,
+		TypeC4Large,
+		TypeC4XLarge,
+		TypeC42XLarge,
+		TypeC44XLarge,
+		TypeC48XLarge,
+		TypeC3Large,
+		TypeC3XLarge,
+		TypeC32XLarge,
+		TypeC34XLarge,
+		TypeC38XLarge,
+		TypeP2XLarge,
+		TypeP28XLarge,
+		TypeP216XLarge,
+		TypeG22XLarge,
+		TypeG28XLarge,
+		TypeX116XLarge,
+		TypeX132XLarge,
+		TypeR4Large,
+		TypeR4XLarge,
+		TypeR42XLarge,
+		TypeR44XLarge,
+		TypeR48XLarge,
+		TypeR416XLarge,
+		TypeR3Large,
+		TypeR3XLarge,
+		TypeR32XLarge,
+		TypeR34XLarge,
+		TypeR38XLarge,
+		TypeI3Large,
+		TypeI3XLarge,
+		TypeI32XLarge,
+		TypeI34XLarge,
+		TypeI38XLarge,
+		TypeI316XLarge,
+		TypeD2XLarge,
+		TypeD22XLarge,
+		TypeD24XLarge,
+		TypeD28XLarge,
+		TypeF12XLarge,
+		TypeF116XLarge,
+	}
+}
