+++ configuration/provider/aws/ec2/instance/default.go
@@ -1,0 +1,7 @@
+package instance
+
+const (
+	// Default is the default instance type used when launching guest clusters and
+	// not specified otherwhise.
+	Default kind = TypeM3Large
+)
